requirements.txt
# Core
pytz==2021.1
Django==3.1.7
django-configurations==2.2
gunicorn==20.0.4
newrelic==6.0.1.155

# For the persistence stores
psycopg2-binary==2.8.6
dj-database-url==0.5.0

# Model Tools
django-model-utils==4.1.1
django_unique_upload==0.2.1

# Rest apis
djangorestframework==3.12.2
Markdown==3.3.3
django-filter==2.4.0

# Developer Tools
ipdb==0.13.4
ipython==7.22.0
mkdocs==1.1.2
flake8==4.0.1
pylint-django==2.5.3

# Testing
mock==4.0.3
factory-boy==3.2.0
django-nose==1.4.7
nose-progressive==1.5.2
coverage==5.4

# Static Files
whitenoise==5.2.0

# ngrok
pyngrok==5.0.5

# Channels (Websockets)
channels==3.0.3
#channels-redis==3.2.0
# using specific working release of channels-redis that fixes lag seen in 3.2.0
#   will be replaced once channels-redis is updated.
git+https://git@github.com/django/channels_redis.git@3656d87813d643bee28fe1960124063268c850a2#egg=channels_redis

# Requests (http requests)
requests==2.25.1

# Constance (app settings)
django-constance[database]==2.8.0
django-extensions==3.1.0

# Websockets
websockets==8.1

django-cors-headers==3.7.0

# Helix Doc Scraping
bs4==0.0.1


# Formatting
autopep8==1.6.0
